Fix backward compatibility issue with QmlProfiler
Previous implementation uses MaxMessage and MaxRangeType constants to
specify events with undefined message or range type. This causes
backwards compatibility issue if new message or range types are
added, because those constants are also written to files when saving
profiler traces.

Add UndefinedMessage and UndefinedRangeType constants and use those
instead of the MaxMessage and MaxRangeType constants. This doesn't fix
opening old traces, but the same problem won't happend again with new
traces.

Also update profiler autotests with fixed data.

Task-number: QTCREATORBUG-28146

+++ src/plugins/qmlprofiler/tests/qmlprofilertraceclient_test.cpp
@@ -77,8 +77,8 @@ void QmlProfilerTraceClientTest::testMessageReceived()
 
         modelManager.replayQmlEvents([&](const QmlEvent &event, const QmlEventType &type) {
             qint64 timestamp;
-            qint32 message;
-            qint32 rangeType;
+            quint8 message;
+            quint8 rangeType;
             checkStream >> timestamp >> message >> rangeType;
             QCOMPARE(event.timestamp(), timestamp);
             QCOMPARE(type.message(), static_cast<Message>(message));
